Rename module to bnscodec
