+++ packages/iov-bns/src/bnscodec.ts
