+++ packages/iov-bns/src/index.ts
@@ -1,2 +1,2 @@
+export { bnsCodec } from "./bnscodec";
 export { Client } from "./client";
-export { bnsCodec } from "./txcodec";
