+++ packages/iov-bns/types/bnscodec.d.ts
