+++ packages/iov-bns/types/index.d.ts
@@ -1,2 +1,2 @@
+export { bnsCodec } from "./bnscodec";
 export { Client } from "./client";
-export { bnsCodec } from "./txcodec";
